added some service tests & logging

+++ proj-1/src/main/java/com/revature/controllers/ItemController.java
@@ -1,6 +1,7 @@
 package com.revature.controllers;
 
 import java.util.List;
+import java.util.UUID;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -24,6 +25,7 @@ import com.revature.exceptions.ItemNotFoundException;
 import com.revature.exceptions.UserNotFoundException;
 import com.revature.models.Item;
 import com.revature.models.User;
+import com.revature.services.AuthService;
 import com.revature.services.ItemService;
 import com.revature.services.UserService;
 
@@ -32,6 +34,7 @@ import com.revature.services.UserService;
 public class ItemController {
 	private static final Logger LOG = LoggerFactory.getLogger(ItemController.class);
 	private ItemService is;
+	private AuthService as;
 
 	@Autowired
 	public ItemController(ItemService is) {
@@ -54,36 +57,57 @@ public class ItemController {
 		return new ResponseEntity<>(is.getAll(),HttpStatus.OK);
 	}
 	
-	
-	
-	@PostMapping
-	public ResponseEntity<String> createItem(@RequestBody Item item/*, @RequestHeader String header*/) {
-		Item i = is.createItem(item);
-		return new ResponseEntity<>("Item " + i.getItemName() + " was created.", HttpStatus.CREATED);
-	}
-	
 	@GetMapping("/{id}")
 	public ResponseEntity<Item> getItemById(@PathVariable("id") int id) {
 		
-			return new ResponseEntity<>(is.getItemById(id), HttpStatus.OK);
+		return new ResponseEntity<>(is.getItemById(id), HttpStatus.OK);
 	}
 
+	
 	@PutMapping("/{id}")
-	public ResponseEntity<Item> updateItem(@RequestBody Item item, @PathVariable("id") int id) {
+	public ResponseEntity<Item> updateItem(@RequestBody Item item, @PathVariable("id") int id, 
+											@RequestHeader(value="Authorization", required=true)String token) {
+		if(token == null) {
+			LOG.warn("Unauthorized user: UPDATE item denied");
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
+		MDC.put("requestId", UUID.randomUUID().toString());
+		as.verify(token);
+		
+		LOG.info("Item id " + id + "updated.");
 		return new ResponseEntity<>(is.updateItem(id, item), HttpStatus.CREATED);
 	}
 	
+	
 	@DeleteMapping("/{id}")
-	public ResponseEntity<String> deleteItem(@PathVariable("id") int id) throws ItemNotFoundException {
+	public ResponseEntity<String> deleteItem(@PathVariable("id") int id, @RequestHeader(value="Authorization", required=true)String token) throws ItemNotFoundException {
+		if (token==null) {
+			LOG.warn("Unauthorized user: DELETE item denied");
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
+//		if (is.getItemById(id) == null) {
+//			return new ItemNotFoundException();
+//		}
+		MDC.put("requestId", UUID.randomUUID().toString());
+		as.verify(token);
+		
 		is.deleteItem(id);
+		LOG.info("Item id " + id + " deleted.");
 		return new ResponseEntity<>("Item deleted", HttpStatus.OK);
 	}
-	
-	
 
-	
-	
-	
-	
+	@PostMapping
+	public ResponseEntity<String> createItem(@RequestBody Item item, @RequestHeader(value = ("Authorization"), required = true)String token) {
+		if (token == null) {
+			LOG.warn("Unauthorized user: CREATE item denied");
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
+		MDC.put("requestId", UUID.randomUUID().toString());
+		as.verify(token);
+		
+		Item i = is.createItem(item);
+		LOG.info("Item id: " + i.getId() + ", Item: " + i.getItemName() + " created.");
+		return new ResponseEntity<>("Item " + i.getItemName() + " was created.", HttpStatus.CREATED);
+	}
 	
 }
